update to symfony 3.0
twig/extensions seems not used
sensio/generator-bundle is for quickstarting
Further an attempt to make the step to symfony 3.0, seems only okish for
php7.0

EntityManager->merge() is to make detached entitie managed again. As the
entity were retrieved with a find(), they should be still managed. So
merge is not needed. (merge is deprecated)

Replace deprecated Twig_SimpleFilter with TwigFilter