autowiring and autoconfigure, uses class names as ids
Type-hinting the constructor arguments let autoload a service, type hint
with interface. See `php bin/console debug:autowiring` for what is
available

Tags in services.yml will let find e.g. twig extensions

Parameters are not autowired, these can be explicit set with
$argumentname: value, the other type-hinted arguments can be left out

Services are default set to private, so autowiring must be used and not
the get('id') of the controller